+++ src/Components/Buttons/Actions/Save.php
@@ -15,11 +15,10 @@ class Save extends SimpleButton
 
         $this->variant ??= 'success';
 
-        $this->confirmVariant ??= 'success';
-
         $this->text ??= Str::ucfirst(trans('action.save'));
 
         if ($this->confirm !== null) {
+            $this->confirmVariant ??= 'success';
             $this->confirmId = 'save-'.($this->confirmId ?? Str::random(32));
         }
     }