+++ CHANGELOG.md
@@ -5,6 +5,10 @@ CHANGELOG
 -------------------
 
 - The `href` attribute of link buttons is not mandatory
+- Added `<x-btn-email />` action button component
+- Added `<x-btn-phone />` action button component
+- Added `<x-btn-website />` action button component
+- Documentation fixes and improvements
 
 
 0.20.1 (2024-10-02)